+++ tamr_client/dataset/dataframe.py
@@ -7,8 +7,8 @@ from typing import Optional
 
 import pandas as pd
 
-from tamr_client.datasets.dataset import Dataset
-import tamr_client.datasets.record as record
+from tamr_client.dataset.dataset import Dataset
+import tamr_client.dataset.record as record
 from tamr_client.session import Session
 from tamr_client.types import JsonDict
